requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Authentication & Security
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Background Tasks
celery==5.3.2
flower==2.0.1

# HTTP & API
requests==2.31.0
marshmallow==3.20.1
webargs==8.3.0

# Notifications
twilio==8.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Monitoring & Logging
prometheus-flask-exporter==0.23.0

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
factory-boy==3.3.0
faker==19.6.2

# Production
gunicorn==21.2.0
eventlet==0.33.3